Renamed some files from blueprint to template
